Moved cookies.py to utils
